Move some of the previously public structs to opaque types.
Experiment with using more idomatic C++ rather than C to interface with TensorRT.
We are using the <class-name>Internal.hpp pattern to expose some of the C++ api to other C++ code.
Not 100% sold on the idea yet but I'm liking how this is starting out. Makes working with the code more familiar and
to the Rust side it should still just appear as a opaque void pointer that we pass when we want to perform an operation.