Add support for ndarray arrays as input and output to execute.
ndarray's are required to be f32's but are generic over the array dimensions. It will return a 1d array that is type f32 for this specific model. Would like to add the ability to specify the return type when creating the engine but that is a feature for the future.