Make the ouptut of context execute generic by taking a mutable reference to an allocated ndarray.
Bump minor version number for tensorrt for the change to the arguments and return type of execute.

I would like to return the result from execute in the future but working with ndarray's as generics has proven to be a bit tricky so this is the best solution for now.