Add build configurations for explicit tensorrt versions.
Add search functionality via ldconfig and grep to make sure we're
finding the correct version of TensorRT installed on the host system.

Cargo features will be used to control what specific libraries we need
to link against to get that version of TensorRT to work correctly.
The features also control the version verification logic to make sure
that the libraries that are found for linking are the versions that we
expect.

The tensorrt version selected via the cargo feature is also passed to
the C++ wrapper code to handle any differences that may exist in the
underlying TensorRT code. Specifically handling deprecations and
removals that happen in the API as the versions change.