Implement `get_binding_dimension` function for the Engine struct.
This allows us to get the dimensions for a specific
binding index and provides an avenue for dynamically
computing the required data size for that binding.

The ssd_uff example is still broken though now
further down in the example. The `execute` function
fails due to an assert inside the nmsPlugin. This
assert fails because we don't have all the outputs
bound for our engine.