Implement multiple output bindings for `execute`
TensorRT engines may have more than one output binding
that needs to be populated. This change takes in outputs
in a Vec of mutable references and populates them in order
that the outputs appear in the model.

To make things a bit more compact we put all of the
bindings (input and output) into one vector assuming that
the first element will always be the input binding and
all others will be the output binding.