Add support for multiple TRT versions
Updated the build system for tensorrt-rs and tensorrt-sys to support bindings to
multiple versions of TensorRT via features. By default tensorrt-rs will create an
API that reflects TensorRT 5. At the time of this commit that can also be changed
to TensorRT 6 or TensorRT 7.

When a feature is selected in tensorrt-rs it will enable the corresponding
feature in tensorrt-sys. This will conditionally include functions in the C++
 bindings that may not have been available in TensorRT 5. At the time of this
 commit there have been no function removals but we can handle that in the future
 if it becomes a problem.

 This commit also switches to using bindgen as a library and excludes the
 generated bindings.rs file from git. bindings.rs can now change depending on the
 features that are enabled/disabled and creates too much noise in the diff.