Test and example error cleanup
Cleaning up some minor build and test issues with the re-worked lifetimes around creating an engine and network.

For the tests in engine.rs moved the Logger in to a lazy static block because
TensorRT expects the logger to be a singleton and will segfault if you pass
multiple instances of builder or runtime a different logger.