Remove un-needed nullptr checks from C++ bindings and cleanup build warnings
I decided to remove the nullptr checks from the builder C++ bindings. I think
they probably caused more issues than they really protected against. Since this
API is not consumed outside of the Rust context which guarantees that the pointer will be valid. If it's not we're in a lot of trouble already and
returning an arbitrary value is kind of pointless.

How to handle this situation is still an open question in my mind. Thinking
maybe terminate the program?