Implement setProfiler for context object
This one is a bit wonky. To implement setProfiler and getProfiler we also had to
create bindings to the `IProfiler` interface. This is expressed in the
`ProfilerBinding` struct which binds to a concrete binding in the C++ wrapper
code that implements the appropriate ILogger interface.

There is a default profiler implemented as `RustProfiler` that will just print
the reported timings to stdout. An end user can implement their own profiler that
does whatever they want with the reported timings via the `IProfiler` trait.