Complete implementation of `nvinfer1::ICudaEngine`
Standard implementaiton of bindings to the `nvinfer1::ICudaEngine` interface.
Only thing to be aware of here is that while the binding to
`createEngineWihtoutDeviceMemory` exists and works. Users of the library won't
be able to use it properly because we don't have support for allocating CUDA
memory on the Rust side yet.