Complete implementation of `nvinfer1::ILayer` base class bindings
All functions available on the `ILayer` base class in TensorRT have been implemented on the
`Layer` trait. This trait is one that can't be impl'd outside of the create since TensorRT
only has a finite amount of layer objects that it supports.

Along with these changes is a generic `BaseLayer` object that represents a plain `ILayer`.
Current thinking is that this will be used to cast into more specific layers.

Also implemented the `IIdentityLayer` which has the same API as the base `ILayer`