Add binding to `addGather` to NetworkDefinition object
Implements the binding for the network definition as well as the `GatherLayer` object
itself.

Also added the tensorrt_rs_derive create to create a custom derive macro for deriving the
`Layer` trait. We're going to be doing this a lot for the next handful of changes so it was
just easier to create a custom derive macro.