Simplify bindings to `ICudaEngine`
Turns out that `bindgen` can actually parse the `NvInfer.h` file
pretty ok in most situations. This removes a lot of the need for
extra indirection and dynamic allocation that we were doing before
to wrap it in a C API.

Only outstanding issue is that it doesn't automatically generate
bindings to pure virtual functions. All member functions in the
`NvInfer.h` file are pure virtual so we just get the types but
not their methods. To fix this we still need a very thin wrapper
that accepts the type as the first parameter and then calls the
appropriate function directly on the type.

There is an issue tracking this
https://github.com/rust-lang/rust-bindgen/issues/27
that hopefully will remove the need for this thin wrapper
completely.