Simplify bindings to various TensorRT structures.
This commit is a bit of a mess. It's mostly a hodge podge of various improvements
and simplifications. Mostly consiting of replacing custom structs with the direct
C++ ones from TensorRT.

Biggest change in this commit is improving the performance of using the various
dims structs. Now we just have a newtype wrapper in Rust that wraps dims instead of
dynamically allocating an opaque pointer in C++. This improves performance where
using dims is needed in a hot code path.