Simplify the TRT Context execute bindings
Focus of this change is two fold. One to simplify the wrapping layer that
connects the Rust library to the C++ library and to gain back some performance
around data input and output when executing a model. I brought in bindings to
cuda itself so we can allocate device side buffers in Rust and pass them to the
C++ wrapper rather than paying for the cost of iterating the buffers twice
before doing the allocation and execution.

Overall these changes provided an 11% improvement in some benchmarks.