Cleanup profiler binding implementation
Minor re-work to the underlying profiler binding implementation to allow us to
effectively sub-class C++ classes from Rust.

Added a `binding_func` macro to abstract the creation of the glue layer that we need
on the Rust side to facilitate passing the call from the subclassed function in C++
to the appropriate function in Rust.